Add: slack status change from timer
